Server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pdfplumber==0.11.4
PyMuPDF==1.25.1
python-docx==1.1.2
